Adds xorg.conf
Not exactly a dotfile, but I'm sick of rewriting this file every
single time I install another Linux distro.

The point is: my monitor native resolution (1366x768) is not
supported by default by my graphics card in Linux. To use this
resolution, I manually had to find on the Internet a valid
Modeline (cvt didn't help as 1366 is not divisible by 8).

Of course, X11 is required for this. I don't know how would Ubuntu
and other "hey-we-are-too-cool-for-X11" distros react to this
files. I don't like Mir either, so that's not a problem to me.