Disable non-breaking spaces on OS X
There is this little "feature" on OS X where pressing Alt+Space
instead of Space will insert a non-breaking space, which is a
character that looks like an Space, but it's a different ASCII
combination (space is 0x20, NBSP is 0xA0).

There are some circunstances in where to type a special character
such as {, | or } I have to press the Option key. Usually I type
so fast that if I have to type a space after that character I
will press Space with one of my hands before my other hand has
released the Option key, making me actually press Option-Space.

This is a problem because many applications like compilers or
shells complain when they find a non-breaking space since they
were expecting... a space. Some examples:

* ls -R | grep ...
  Error: the character after the pipe is not a space -- is a NBSP.
* } else {
  Error: the character after } is not a space -- is a NBSP.

This hack don't work with some applications, but at least it works
on Terminal.app, where I don't usually have cues that tell me if
I added an NBSP. Some applications like Eclipse or NetBeans will
underline the NBSP as I'm writing it so I can notice. This doesn't
work on MacVim out of the box.