Add private profiles to PowerShell
Say that I want to have private profiles that are only available
for a particular machine given its name. I don't want these scripts
to run on other machines.

As an experimental test, I'm adding private profiles to PowerShell
and planning on adding them to Bash, Zsh and Vim as well, where
if a private profile exists, it will be sourced as well.

Private profiles for PowerShell should be named using the name of
the machine, ending with .ps1 for PowerShell scripts. As an example,
C:\Users\danirod\Documents\WindowsPowerShell\MACHINE-2.ps1 will be
only sourced on machines whose name is MACHINE-2.ps1.

These private profiles can be added to Git or excluded from Git,
allowing me to put here code that might be confidential or not visible
to outsiders, without having to blacklist my main profile file or
having to manually exclude snippets every time I commit, which I
would eventually fail.