Clean up fonts.conf syntax
Oh yeah, finally got this. Original fonts.conf file as introduced in
commit 25c2247 had more matches than what it should have. Since all the
edits apply all the time, they could be simplified into a single match
rule.

Plus this will make it easier to me to override the full hinting for
particular font families because, you know, the original version comes
from 08-2013, but this is 2017 and every web site has to use a shitty
webfont that looks ugly when full hinting is enabled.