[nvim] Upgrade tagbar settings
This commit removes the secondary keybinding for opening tagbar. <F8>
will now open instead of toggle tagbar, and it will focus the tagbar
window for a faster access to the outline.

Also, the width of the tagbar window is now dynamic and it depends on
how big is the screen, so that it takes more screen state in a 1920x1080
display than on a 1280x800 or a 1366x768 display. If the window changes,
the size of the pane will update, but I don't know how unefficient this
is.