[nvim] Configure syntastic
If a file named .syntastic_c_config is found in the project, these
settings will be passed to GCC or clang when syntastic lints the C code.

I added this setting because Syntastic was marking all the include
statements in my kernel code as invalid because Syntastic didn't know
that I have a custom directory added in my include path.