[bash] better BROWSER detector
Unifies the differences between Darwin and Linux in terms of browser
detection. Still using `open` as the default browser for macOS, as it
will open the browser set on system settings. On Linux, the browser is
now set by looking at more alternatives, for use on workstations
(Firefox), going all the way down to Raspberries (Dillo).